Rename contact route to sendMail for more general use.
