+++ app.js
@@ -11,7 +11,7 @@ var homepage = require('./routes/homepage')
 var about = require('./routes/about')
 var blogs = require('./routes/blogs')
 var news = require('./routes/news')
-var contact = require('./routes/contact')
+var sendMail = require('./routes/sendMail')
 var faqs = require('./routes/faqs')
 var privacypolicy = require('./routes/privacypolicy')
 var careers = require('./routes/careers')
@@ -36,7 +36,7 @@ app.use('/homepage', homepage)
 app.use('/about', about)
 app.use('/blogs', blogs)
 app.use('/news', news)
-app.use('/contact', contact)
+app.use('/sendMail', sendMail)
 app.use('/faqs', faqs)
 app.use('/privacypolicy', privacypolicy)
 app.use('/careers', careers)
